Nexus/requirements.txt
dj-user-login-history
django-admin-logs
django-image-uploader-widget
python-decouple
django-jazzmin
django-defender
django-embed-video
django-widget-tweaks
django-auto-logout
pytz
django-daisy  #this is admin pannel template